+++ src/main/java/it/pagopa/pn/deliverypush/actions/NotificationViewedActionHandler.java
@@ -1,5 +1,6 @@
 package it.pagopa.pn.deliverypush.actions;
 
+import it.pagopa.pn.deliverypush.legalfacts.LegalFactUtils;
 import org.springframework.stereotype.Component;
 
 import it.pagopa.pn.api.dto.notification.Notification;
