+++ src/main/java/it/pagopa/pn/deliverypush/actions/SenderAckActionHandler.java
@@ -3,6 +3,7 @@ package it.pagopa.pn.deliverypush.actions;
 import java.time.Instant;
 import java.util.stream.Collectors;
 
+import it.pagopa.pn.deliverypush.legalfacts.LegalFactUtils;
 import org.springframework.stereotype.Component;
 
 import it.pagopa.pn.api.dto.notification.Notification;
