+++ src/main/resources/application.properties
@@ -4,6 +4,7 @@ pn.env.runtime=PROD
 # Application parameters
 pn.middleware.impl.delivery-dao=cassandra
 pn.middleware.impl.timeline-dao=cassandra
+pn.legalfacts.generator=OPENHTML2PDF
 
 # 15 minutes
 pn.delivery-push.webhook.schedule-interval=900000
