+++ src/main/resources/openhtmltopdf/openHtmlToPdf.html
@@ -8,7 +8,7 @@
 	}
 
 	.paragraph {
-		font-family: "Times New Roman", Times, serif;
+		font-family: Courier, serif;
 		font-size: 16px;
 		margin-top: 25px;
 	}
